saead: log block length with %u instead of %zu
The plaintext/ciphertext length locals are uint16_t, so %zu read a
size_t from an int-promoted argument, which is undefined behavior and a
-Wformat mismatch.  Print them with %u and an explicit cast, matching
the style used elsewhere for these length fields.

+++ src/saead/session.c
@@ -225,7 +225,7 @@ struct pouch_buf *session_encrypt_block(struct session *session, struct pouch_bu
 
     if (plaintext_len != pouch_bufview_available(&plaintext))
     {
-        POUCH_LOG_ERR("Invalid plaintext length: %zu", plaintext_len);
+        POUCH_LOG_ERR("Invalid plaintext length: %u", (unsigned int) plaintext_len);
         buf_free(encrypted);
         return NULL;
     }
@@ -293,7 +293,7 @@ int session_decrypt_block(struct session *session,
 
     if (ciphertext_len <= AUTH_TAG_LEN || ciphertext_len != pouch_bufview_available(&ciphertext))
     {
-        POUCH_LOG_ERR("Invalid ciphertext length: %zu", ciphertext_len);
+        POUCH_LOG_ERR("Invalid ciphertext length: %u", (unsigned int) ciphertext_len);
         return -EINVAL;
     }
